… btl/resources/shapes/_fusion_ball_end_mill.fcstd
… btl/resources/shapes/_fusion_ball_end_mill.png
… btl/resources/shapes/_fusion_bull_nose_end_mill.fcstd
… btl/resources/shapes/_fusion_bull_nose_end_mill.png
… btl/resources/shapes/_fusion_center_drill.fcstd
… btl/resources/shapes/_fusion_center_drill.png
… btl/resources/shapes/_fusion_chamfer_mill.fcstd
… btl/resources/shapes/_fusion_chamfer_mill.png
… btl/resources/shapes/_fusion_counter_sink.fcstd
… btl/resources/shapes/_fusion_counter_sink.png
… btl/resources/shapes/_fusion_dovetail_mill.fcstd
… btl/resources/shapes/_fusion_dovetail_mill.png
… btl/resources/shapes/_fusion_drill.fcstd
… btl/resources/shapes/_fusion_drill.png
… btl/resources/shapes/_fusion_face_mill.fcstd
… btl/resources/shapes/_fusion_face_mill.png
… btl/resources/shapes/_fusion_flat_end_mill.fcstd
… btl/resources/shapes/_fusion_flat_end_mill.png
… btl/resources/shapes/_fusion_lollipop_mill.fcstd
… btl/resources/shapes/_fusion_lollipop_mill.png
… btl/resources/shapes/_fusion_radius_mill.fcstd
… btl/resources/shapes/_fusion_radius_mill.png
… btl/resources/shapes/_fusion_reamer.fcstd
… btl/resources/shapes/_fusion_reamer.png
… btl/resources/shapes/_fusion_slot_mill.fcstd
… btl/resources/shapes/_fusion_slot_mill.png
… btl/resources/shapes/_fusion_tap.fcstd
… btl/resources/shapes/_fusion_tap.png
… btl/resources/shapes/_fusion_tapered_mill.fcstd
… btl/resources/shapes/_fusion_tapered_mill.png