closes #7

Example data to test this is available at [https://cam.autodesk.com/hsmtools](https://cam.autodesk.com/hsmtools)

I added custom tool shapes for importing from fusion, because the dimensions and shapes don't translate 1:1 E.G.:
- fusion specifies an additional shoulder length parameter for all tools that isn't present in most of the btl shapes. 
- dovetail cutter have rounded edges instead of truncated
- some tools may have rounded or square edges, so the shape .fcstd file need to swap the revolved sketch out on the fly

Perhaps we can figure out a way to hide these import specific shapes from the btl add tool gui?

![image](https://github.com/knipknap/better-tool-library/assets/37948669/646fa485-6d91-4ea8-a096-e931c78d7a09)

There is also a performance issue somewhere - a 15Mb json library takes ~5 minutes to import.

This is definitely still in a draft state. Have a look, see what you think.