@knipknap 2024-01-10
Wow, that looks crazy awesome, great work!

Hopefully the performance issue is not caused by the renderer, because that could be hard to fix.