@alexneufeld 2024-01-11
> Wow, that looks crazy awesome, great work!
> 
> Hopefully the performance issue is not caused by the renderer, because that could be hard to fix.

Doesn't look like the renderer is the issue - `btl/library.py` just isn't efficient enough to handle ~1000 tools. Here are my preliminary optimization results:

test data: [https://cam.autodesk.com/tools/download.php?name=Harvey%20Tool/End%20Mills](https://cam.autodesk.com/tools/download.php?name=Harvey%20Tool/End%20Mills)

test script:

``` python
# run with freecad -c ~/test_script.py >> optimization_logs.txt
import os
import sys
from cProfile import Profile
from pstats import SortKey, Stats
from btl.serializers import FusionJSONSerializer


with Profile() as profile:
    filename = "/path/to/Harvey Tool-End Mills.json"
    dirname = os.path.dirname(filename)
    serializer = FusionJSONSerializer(dirname)
    # Try to read the file.
    library = serializer.deserialize_library_from_file(filename)
    (
        Stats(profile)
        .strip_dirs()
        .sort_stats(SortKey.CALLS)
        .print_stats()
    )

sys.exit(0)

```

trimmed test results:

```
// before optimization
363418835 function calls (363406722 primitive calls) in 123.958 seconds

   Ordered by: call count

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
356188145   62.841    0.000   62.841    0.000 tool.py:34(__eq__)
  1103131    0.164    0.000    0.164    0.000 {built-in method builtins.isinstance}
   493808    0.065    0.000    0.065    0.000 {method 'get' of 'dict' objects}
   304032    0.052    0.000    0.052    0.000 {method 'join' of 'str' objects}
   302660    0.047    0.000    0.047    0.000 {method 'group' of 're.Match' objects}
   302656    0.437    0.000    0.536    0.000 sre_parse.py:1066(expand_template)
   302656    0.116    0.000    0.653    0.000 re.py:330(filter)
   223018    0.033    0.000    0.033    0.000 {built-in method sys._getframe}
223018/222967    0.099    0.000    0.538    0.000 {built-in method builtins.__import__}
223015/222967    0.426    0.000    1.042    0.000 feature.py:105(_import)
...

// after optimization

7111326 function calls (7111149 primitive calls) in 6.957 seconds

   Ordered by: call count

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
  1103131    0.154    0.000    0.154    0.000 {built-in method builtins.isinstance}
   481871    0.054    0.000    0.054    0.000 {method 'get' of 'dict' objects}
   304032    0.050    0.000    0.050    0.000 {method 'join' of 'str' objects}
   302660    0.044    0.000    0.044    0.000 {method 'group' of 're.Match' objects}
   302656    0.428    0.000    0.521    0.000 sre_parse.py:1066(expand_template)
   302656    0.119    0.000    0.640    0.000 re.py:330(filter)
   223018    0.030    0.000    0.030    0.000 {built-in method sys._getframe}
223018/222967    0.093    0.000    0.556    0.000 {built-in method builtins.__import__}
223015/222967    0.414    0.000    1.044    0.000 feature.py:105(_import)
...

```

df5502322bbb6c771510d256b1cfde95b969b1a3 is pretty unsophisticated, I just deleted some expensive safety checks in the tool addition process. Looks like storing the tools as a list is particularly bad because of how often we check for membership in the library and or iterate through it:

``` python
class Library(object):
    API_VERSION = 1

    def __init__(self, label, id=None):
        self.id = id or str(uuid.uuid1())
        self.label = label
        self.tools = []  # <-- bad
        ...
```

Any thoughts?