@knipknap 2024-01-11
That looks like quite the performance improvement! The tools are already stored in two data structures to improve the performance of lookups, the next line in the constructor is:

        self.tool_nos = {}  # Maps tool_no number to tool

But neither structure allows for easily finding the next available tool number.

When implementing the "assign_new_tool_number" I was trying to avoid using the highest number, because that used to cause some inconvenience in the FreeCAD internal tool library UI before. Some logic was added to avoid "accidental reshuffling" when passing by an already occupied tool number while a user spins through them in the UI, using the spin buttons of the tool number field.

But I think my reasoning has become rather obsolete, since after introducing the pocket attribute, the situations where "just using the highest number" was a problem should no longer exist. At least I can't think of one anymore.

So in short, I think using a tool_no_inc attribute is good.

Just one small nitpick: We have to be very careful with the naming, as it's easy to confuse tool ID, tool number, and tool pocket. So I'd suggest to call the counter tool_no_inc or something.